(refactor(repo-automation)): extract workflow registration helpers
- Move interactive repo-automation command registration out of extension.ts to restore the repository file-size budget without changing command behavior or IDs
- Extract codex-native converter, prompt-resolution, and orchestration helper builders from repo-automation-service.ts behind the existing service contract
- Add focused workflow command coverage for the extracted paths and stage the remediation audit, feature review, and QA evidence for issue #164

Refs: #164
