+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/phase0-instructions-read.2026-04-26T19-20.md
@@ -1,0 +1,11 @@
+Timestamp: 2026-04-26T19-20
+Command: document-policy-reads
+EXIT_CODE: 0
+Output Summary: Phase 0 policy files read in full-feature order.
+Policy Order: 1..5
+Files Read:
+- .github/copilot-instructions.md
+- .github/instructions/general-code-change.instructions.md
+- .github/instructions/general-unit-test.instructions.md
+- .github/instructions/typescript-code-change.instructions.md
+- .github/instructions/typescript-unit-test.instructions.md
