+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-repo-automation-service-lines.2026-04-26T19-20.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19-20
+Command: pwsh -NoProfile -Command "(Get-Content 'extensions/drm-copilot/src/repo-automation-service.ts' | Measure-Object -Line).Lines"
+EXIT_CODE: 0
+Output Summary: 560
