+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-scope.2026-04-26T19-20.md
@@ -1,0 +1,9 @@
+Timestamp: 2026-04-26T19-20
+Command: document-remediation-scope
+EXIT_CODE: 0
+Output Summary: Acceptance criteria remain PASS; remediation scope limited to R-1 through R-3.
+Acceptance Criteria Reopened: no
+Findings In Scope:
+- R-1: Reduce extensions/drm-copilot/src/extension.ts to 500 lines or fewer without changing public command behavior.
+- R-2: Reduce extensions/drm-copilot/src/repo-automation-service.ts to 500 lines or fewer without changing the repo-automation service contract.
+- R-3: Regenerate review artifacts against the correct working-tree scope after remediation and confirm the policy findings are closed.
