+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-typescript-format-check.2026-04-26T19-20.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19-20
+Command: npm --prefix extensions/drm-copilot exec prettier --check "src/**/*.ts" "test/**/*.ts" "*.json" "*.cjs"
+EXIT_CODE: 0
+Output Summary: The baseline check invocation was not a clean formatting status signal on Windows. npm emitted an Unknown cli config warning for --check and Prettier reported no files matching the requested patterns, including test/**/*.ts. The current baseline therefore needs the later final format gate to provide the authoritative formatting result.
