+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-typescript-lint.2026-04-26T19-20.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19-20
+Command: npm --prefix extensions/drm-copilot run lint
+EXIT_CODE: 0
+Output Summary: Pass. ESLint completed with no reported diagnostics for src and test.
