+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-typescript-test-coverage.2026-04-26T19-20.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19-20
+Command: npm --prefix extensions/drm-copilot run test:unit -- --coverage
+EXIT_CODE: 0
+Output Summary: Pass. Jest ran 32 suites and 345 tests successfully. Coverage headlines: All files lines 94.42%, extension.ts lines 90.67%, repo-automation-service.ts lines 100.00%.
