+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-extension-lines-after.2026-04-26T19-20.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19-20
+Command: pwsh -NoProfile -Command "(Get-Content 'extensions/drm-copilot/src/extension.ts' | Measure-Object -Line).Lines"
+EXIT_CODE: 0
+Output Summary: 268
