+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-repo-automation-service-split-boundary.2026-04-26T19-20.md
@@ -1,0 +1,8 @@
+Timestamp: 2026-04-26T19-20
+Command: document-repo-automation-service-split-boundary
+EXIT_CODE: 0
+Output Summary: Identified helper-option construction and asset-resolution logic as the extraction boundary for repo-automation-service.ts.
+Boundary Notes:
+- Keep RepoAutomationService interfaces, DefaultRepoAutomationService, executeScript(), and runPoshQcWorkflow() in repo-automation-service.ts so the public orchestration boundary remains stable.
+- Extract the codex-native converter, policy-audit asset resolution, execute-hard-lock prompt, atomic-plan prompt, validation, and adjacent argument-building logic into focused helper functions in a separate module.
+- Preserve RunCodexNativeConverterInput, public method names, tool names, summaries, and returned artifact semantics exactly.
