+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-final-verdict.2026-04-26T19-20.md
@@ -1,0 +1,35 @@
+# Remediation Final Verdict
+
+Timestamp: 2026-04-26T23:43:00
+Command: document-remediation-final-verdict
+EXIT_CODE: 0
+Output Summary: Structural remediation closed the two TypeScript file-size blockers, the final TypeScript QA loop passed, and the rerun review artifacts validated successfully.
+Verdict: go
+
+## Post-Remediation Line Counts
+
+- `extensions/drm-copilot/src/extension.ts`: 268 lines
+- `extensions/drm-copilot/src/repo-automation-service.ts`: 473 lines
+
+## Final TypeScript QA Results
+
+- `npm --prefix extensions/drm-copilot run format` → PASS
+- `npm --prefix extensions/drm-copilot run lint` → PASS
+- `npm --prefix extensions/drm-copilot run typecheck` → PASS
+- `npm --prefix extensions/drm-copilot run test:unit -- --coverage` → PASS
+- Coverage highlights from the final run:
+  - `extension.ts`: 98.63%
+  - `repo-automation-service.ts`: 100.00%
+  - `repo-automation-command-registration.ts`: 98.55%
+  - `repo-automation-service-workflows.ts`: 100.00%
+
+## Review Artifact Rerun Outcomes
+
+- `policy-audit.2026-04-26T23-41.md` regenerated and validated successfully.
+- `code-review.2026-04-26T23-41.md` regenerated and validated successfully.
+- `feature-audit.2026-04-26T23-41.md` regenerated and validated successfully.
+- The rerun artifacts report no remaining FAIL findings in the remediation scope.
+
+## Residual Blockers
+
+None. The branch is ready for re-review against the refreshed audit set.
