+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-pr-context-status.2026-04-26T19-20.md
@@ -1,0 +1,8 @@
+Timestamp: 2026-04-26T19-20
+Command: poetry run python -m scripts.dev_tools.pr_context.collector --base development
+EXIT_CODE: 0
+Output Summary: Refreshed PR context artifacts successfully. The post-remediation review scope now comes from a non-empty commit range against development rather than the working-tree appendix.
+Scope Basis: commit range
+Base Ref: origin/development @ 0762f58a1451994999c2f49f2dbdc489120d138a
+Head Ref: feature/codex-native-converter-164 @ 79d02b78bfe52a0457a3a899e0fca7a2e03a4adb
+Range: 0762f58a1451994999c2f49f2dbdc489120d138a..79d02b78bfe52a0457a3a899e0fca7a2e03a4adb
