+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-typescript-test-coverage.2026-04-26T19-20.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19-20
+Command: npm --prefix extensions/drm-copilot run test:unit -- --coverage
+EXIT_CODE: 0
+Output Summary: Pass. Jest ran 32 suites and 349 tests successfully. Coverage headlines: All files lines 95.49%, extension.ts lines 98.63%, repo-automation-service.ts lines 100.00%, repo-automation-command-registration.ts lines 98.55%, repo-automation-service-workflows.ts lines 100.00%. Changed-file coverage remains at or above 90% for the touched production files.
