+++ docs/features/active/2026-04-26-codex-native-converter-164/feature-audit.2026-04-26T23-41.md
@@ -1,0 +1,30 @@
+# Feature Audit: codex-native-converter (#164)
+
+## Scope and Baseline
+
+This rerun reviews the remediation for findings R-1 through R-3 only. Baseline line counts were 687 for `extensions/drm-copilot/src/extension.ts` and 560 for `extensions/drm-copilot/src/repo-automation-service.ts`; the post-remediation counts are 268 and 473 respectively. The final TypeScript QA evidence comes from the restarted clean pass recorded under `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/`.
+
+## Acceptance Criteria Inventory
+
+- Source files: `docs/features/active/2026-04-26-codex-native-converter-164/spec.md` and `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`
+- Original feature acceptance criteria were already satisfied before remediation and were not reopened by this structural fix.
+- Remediation-specific success conditions were: restore the 500-line structural limit, preserve existing command and service behavior, rerun the TypeScript QA loop, and refresh review artifacts.
+
+## Acceptance Criteria Evaluation
+
+| Item | Status | Evidence | Notes |
+| --- | --- | --- | --- |
+| Original user-story acceptance criteria remain satisfied | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`, prior `feature-audit.2026-04-26T19-20.md` | The remediation did not reopen feature behavior. |
+| `extension.ts` and `repo-automation-service.ts` comply with the 500-line limit | PASS | `remediation-extension-lines-after.2026-04-26T19-20.md`, `remediation-repo-automation-service-lines-after.2026-04-26T19-20.md` | Both touched production files are now below 500 lines. |
+| Final TypeScript QA loop passes with required coverage evidence | PASS | `remediation-typescript-format.2026-04-26T19-20.md`, `remediation-typescript-lint.2026-04-26T19-20.md`, `remediation-typescript-typecheck.2026-04-26T19-20.md`, `remediation-typescript-test-coverage.2026-04-26T19-20.md` | Final clean pass completed after the required loop restarts. |
+| Review rerun basis is explicit | PASS | `remediation-pr-context-status.2026-04-26T19-20.md`, `artifacts/pr_context.summary.txt` | The rerun now uses a non-empty commit range against `development`. |
+
+## Summary
+
+**Overall Feature Readiness:** PASS
+
+The remediation closed the only remaining review blocker without reopening delivered feature behavior. The branch is structurally compliant in the remediated scope and is ready for normal review follow-up.
+
+## Acceptance Criteria Check-off
+
+No additional acceptance-criteria checkboxes were changed during remediation because the original feature criteria were already PASS before the structural fix. The remediation preserved that PASS state while closing the policy blocker.
