+++ docs/features/active/2026-04-26-codex-native-converter-164/policy-audit.2026-04-26T23-41.md
@@ -1,0 +1,82 @@
+# Policy Audit: codex-native-converter remediation (#164)
+
+**Reviewer:** atomic_executor (GitHub Copilot)
+**Review scope:** Post-remediation branch state for `feature/codex-native-converter-164` against `development`
+**Code Under Test:** `extensions/drm-copilot/src/extension.ts`, `extensions/drm-copilot/src/repo-automation-service.ts`, `extensions/drm-copilot/src/repo-automation-command-registration.ts`, `extensions/drm-copilot/src/repo-automation-service-workflows.ts`, `extensions/drm-copilot/test/extension.workflow-commands.test.ts`, and remediation evidence under `docs/features/active/2026-04-26-codex-native-converter-164/evidence/`
+
+## Executive Summary
+
+The remediation closed the structural policy blocker by reducing the two oversized touched TypeScript production files below the repository 500-line limit while preserving command behavior and the repo-automation service contract. The final TypeScript QA loop completed with a clean pass: formatting, linting, type checking, and Jest coverage all passed. The refreshed PR context now resolves to a non-empty commit range against `development`, so the rerun review scope is explicit.
+
+## 1. General Unit Test Policy Compliance
+
+- TypeScript baseline coverage artifact: `docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-typescript-test-coverage.2026-04-26T19-20.md`
+- TypeScript post-change coverage artifact: `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-typescript-test-coverage.2026-04-26T19-20.md`
+- PowerShell baseline coverage artifact: N/A - remediation touched no PowerShell files.
+- PowerShell post-change coverage artifact: N/A - remediation touched no PowerShell files.
+- Per-language comparison summary: TypeScript baseline 94.42%, post-change 95.49%, new or changed production files 98.55% or higher.
+
+### 1.2.1 Per-Language Coverage Comparison
+
+- TypeScript: Baseline: 94.42%; Post-change: 95.49%; Change: +1.07%; New/changed-code coverage: 98.55%; Disposition: PASS; Evidence: `docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-typescript-test-coverage.2026-04-26T19-20.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-typescript-test-coverage.2026-04-26T19-20.md`
+
+## 2. General Code Change Policy Compliance
+
+The remediation stayed within findings R-1 through R-3, preserved public command identifiers, and restored structural compliance by splitting the oversized TypeScript modules into focused helpers. Post-remediation line counts are `268` for `extension.ts` and `473` for `repo-automation-service.ts`.
+
+## 3. Language-Specific Code Change Policy Compliance
+
+TypeScript changes remain within the approved extension package surface, use the repository toolchain, and preserve the thin-wrapper boundary over the bundled Python converter. The extracted modules keep `extension.ts` as an activation coordinator and keep `repo-automation-service.ts` as the stable service contract.
+
+## 4. Language-Specific Unit Test Policy Compliance
+
+Focused Jest coverage was added for the newly extracted registration module, and the final package-wide run passed 32 suites and 349 tests. New production helper modules are above the repository 90% coverage target.
+
+## 5. Test Coverage Detail
+
+| Language | Files Changed | Tests | Test Result | Baseline Coverage | Post-Change Coverage | New Code Coverage |
+| --- | --- | --- | --- | --- | --- | --- |
+| TypeScript | `extension.ts`, `repo-automation-service.ts`, `repo-automation-command-registration.ts`, `repo-automation-service-workflows.ts`, `extension.workflow-commands.test.ts` | Jest unit tests with coverage | PASS | 94.42% | 95.49% | 98.55% |
+
+## 6. Test Execution Metrics
+
+- Formatter: `npm --prefix extensions/drm-copilot run format` → PASS on the final clean pass.
+- Linter: `npm --prefix extensions/drm-copilot run lint` → PASS.
+- Type checker: `npm --prefix extensions/drm-copilot run typecheck` → PASS.
+- Test runner: `npm --prefix extensions/drm-copilot run test:unit -- --coverage` → PASS, 32 suites, 349 tests.
+
+## 7. Code Quality Checks
+
+- Structural limit restored: `extension.ts` 687 → 268 lines.
+- Structural limit restored: `repo-automation-service.ts` 560 → 473 lines.
+- PR context refreshed: `artifacts/pr_context.summary.txt` now records a non-empty commit range to `origin/development`.
+
+## 8. Gaps and Exceptions
+
+No open policy gaps remain in the remediation scope. The baseline formatting-check command recorded in Phase 0 behaved oddly on Windows through `npm exec`, but the final repository-approved formatting command completed cleanly and serves as the authoritative formatting result.
+
+## 9. Summary of Changes
+
+The remediation introduced `repo-automation-command-registration.ts` and `repo-automation-service-workflows.ts`, reduced the two oversize files below the repository limit, added focused workflow tests for the extracted registration paths, refreshed PR context, and produced a clean final TypeScript QA pass.
+
+## 10. Compliance Verdict
+
+PASS. The structural blocker identified in the prior review is closed, the touched production files are now below the 500-line limit, and the final TypeScript QA loop passed with numeric coverage evidence.
+
+## Appendix A: Test Inventory
+
+- `test/extension.test.ts`
+- `test/mcp-tools.codex-native-converter.test.ts`
+- `test/repo-automation-service.test.ts`
+- `test/repo-automation-service.codex-native-converter.test.ts`
+- `test/repo-automation-dispatch.test.ts`
+- `test/extension.workflow-commands.test.ts`
+- Package-wide Jest coverage run via `npm --prefix extensions/drm-copilot run test:unit -- --coverage`
+
+## Appendix B: Toolchain Commands Reference
+
+- `npm --prefix extensions/drm-copilot run format`
+- `npm --prefix extensions/drm-copilot run lint`
+- `npm --prefix extensions/drm-copilot run typecheck`
+- `npm --prefix extensions/drm-copilot run test:unit -- --coverage`
+- `poetry run python -m scripts.dev_tools.pr_context.collector --base development`
