+++ extensions/drm-copilot/src/repo-automation-service.ts
@@ -4,21 +4,21 @@ import {
   executeBundledScriptFromExtensionRoot,
 } from "./command-runtime";
 import {
-  copyBundledPolicyAuditTemplateAsset,
-  resolveBundledPolicyAuditTemplateAsset,
-} from "./policy-audit-template-assets";
-import {
   normalizeGeneratedPath,
   parseFirstArtifactPath,
   type ScriptExecutionOptions,
 } from "./repo-automation-service-support";
 import { type RepoAutomationToolName } from "./repo-automation-tool-names";
+import { buildPoshQcWorkflowArguments } from "./repo-automation-args";
 import {
-  buildNewActiveFeatureFolderArgs,
-  buildPoshQcWorkflowArguments,
-  buildResolveExecuteHardLockPromptArguments,
-  buildValidateOrchestrationArtifactsArgs,
-} from "./repo-automation-args";
+  buildNewActiveFeatureFolderOptions,
+  buildResolveAtomicPlanPromptOptions,
+  buildResolveExecuteHardLockPromptOptions,
+  buildRunCodexNativeConverterOptions,
+  buildTemplateRoot,
+  buildValidateOrchestrationArtifactsOptions,
+  resolvePolicyAuditTemplateAssetResult,
+} from "./repo-automation-service-workflows";
 import {
   type PolicyAuditTemplateAssetSelector,
   type PotentialPromotionType,
@@ -159,9 +159,7 @@ class DefaultRepoAutomationService implements RepoAutomationService {
   constructor(options: RepoAutomationServiceOptions) {
     this.extensionRoot = options.extensionRoot;
     this.output = options.output;
-    this.templateRoot = normalizeGeneratedPath(
-      path.join(this.extensionRoot, "resources", "feature-templates"),
-    );
+    this.templateRoot = buildTemplateRoot(this.extensionRoot);
   }
   async collectCommitContext(
     input: WorkspaceExecutionInput,
@@ -214,40 +212,7 @@ class DefaultRepoAutomationService implements RepoAutomationService {
   async runCodexNativeConverter(
     input: RunCodexNativeConverterInput,
   ): Promise<RepoAutomationExecutionResult> {
-    const args = [
-      input.mode,
-      "--source-root",
-      input.sourceRoot,
-      "--source-ecosystem",
-      input.sourceEcosystem,
-    ];
-
-    if (input.destinationRoot !== undefined) {
-      args.push("--destination-root", input.destinationRoot);
-    }
-
-    if (input.artifactRoot !== undefined) {
-      args.push("--artifact-root", input.artifactRoot);
-    }
-
-    if (input.enableRepoPrompts === true) {
-      args.push("--enable-repo-prompts");
-    }
-
-    for (const selectedPath of input.selectedPaths ?? []) {
-      args.push("--selected-path", selectedPath);
-    }
-
-    return this.executeScript({
-      tool: "run_codex_native_converter",
-      runtimeKind: "python",
-      bundledRelativePath: "resources/templates/codex_native_converter.py",
-      workspaceRoot: input.workspaceRoot,
-      invocationId: input.invocationId ?? "run_codex_native_converter",
-      args,
-      summary: `Ran bundled codex-native-converter in ${input.mode} mode for '${input.sourceEcosystem}'.`,
-      stdoutArtifactPattern: /Artifact root:\s*(.+)/i,
-    });
+    return this.executeScript(buildRunCodexNativeConverterOptions(input));
   }
   async pushDownCopilotCustomizations(
     input: WorkspaceExecutionInput,
@@ -383,16 +348,9 @@ class DefaultRepoAutomationService implements RepoAutomationService {
       readonly workMode: WorkModeOption;
     },
   ): Promise<RepoAutomationExecutionResult> {
-    const args = buildNewActiveFeatureFolderArgs(input, this.templateRoot);
-    return this.executeScript({
-      tool: "new_active_feature_folder",
-      runtimeKind: "python",
-      bundledRelativePath: "resources/templates/new_active_feature_folder.py",
-      workspaceRoot: input.workspaceRoot,
-      invocationId: input.invocationId ?? "new_active_feature_folder",
-      args,
-      summary: `Created a new active ${input.type} feature folder for '${input.featureName}'.`,
-    });
+    return this.executeScript(
+      buildNewActiveFeatureFolderOptions(input, this.templateRoot),
+    );
   }
   async runPoshQCFormat(
     input: WorkspaceExecutionInput & {
@@ -436,33 +394,7 @@ class DefaultRepoAutomationService implements RepoAutomationService {
       readonly targetPath?: string;
     },
   ): Promise<RepoAutomationExecutionResult> {
-    const resolvedAsset = resolveBundledPolicyAuditTemplateAsset(
-      this.extensionRoot,
-      input.asset,
-    );
-    const destinationPath =
-      input.targetPath === undefined
-        ? undefined
-        : copyBundledPolicyAuditTemplateAsset(
-            resolvedAsset.bundledSourcePath,
-            input.targetPath,
-          );
-
-    return {
-      tool: "resolve_policy_audit_template_asset",
-      workspaceRoot: input.workspaceRoot,
-      summary:
-        destinationPath === undefined
-          ? `Resolved bundled policy-audit asset '${input.asset}'.`
-          : `Copied bundled policy-audit asset '${input.asset}' to '${destinationPath}'.`,
-      artifacts:
-        destinationPath === undefined
-          ? [resolvedAsset.bundledSourcePath]
-          : [resolvedAsset.bundledSourcePath, destinationPath],
-      assetId: resolvedAsset.assetId,
-      bundledSourcePath: resolvedAsset.bundledSourcePath,
-      ...(destinationPath === undefined ? {} : { destinationPath }),
-    };
+    return resolvePolicyAuditTemplateAssetResult(this.extensionRoot, input);
   }
 
   async resolveExecuteHardLockPrompt(
@@ -472,33 +404,13 @@ class DefaultRepoAutomationService implements RepoAutomationService {
       readonly quiet?: boolean;
     },
   ): Promise<RepoAutomationExecutionResult> {
-    const { args, artifactPaths } =
-      buildResolveExecuteHardLockPromptArguments(input);
-
-    return this.executeScript({
-      tool: "resolve_execute_hard_lock_prompt",
-      runtimeKind: "python",
-      bundledRelativePath: "resources/templates/resolve_hard_lock_prompt.py",
-      workspaceRoot: input.workspaceRoot,
-      invocationId: input.invocationId ?? "resolve_execute_hard_lock_prompt",
-      args,
-      summary: `Resolved the execute hard-lock prompt for '${input.target}'.`,
-      ...(artifactPaths === undefined ? {} : { artifactPaths }),
-    });
+    return this.executeScript(buildResolveExecuteHardLockPromptOptions(input));
   }
 
   async resolveAtomicPlanPrompt(
     input: WorkspaceExecutionInput & { readonly target: string },
   ): Promise<RepoAutomationExecutionResult> {
-    return this.executeScript({
-      tool: "resolve_atomic_plan_prompt",
-      runtimeKind: "python",
-      bundledRelativePath: "resources/templates/resolve_atomic_plan_prompt.py",
-      workspaceRoot: input.workspaceRoot,
-      invocationId: input.invocationId ?? "resolve_atomic_plan_prompt",
-      args: ["--target", input.target, "--workspace", input.workspaceRoot],
-      summary: `Resolved the atomic-plan prompt for '${input.target}'.`,
-    });
+    return this.executeScript(buildResolveAtomicPlanPromptOptions(input));
   }
 
   private async runPoshQcWorkflow(
@@ -535,18 +447,9 @@ class DefaultRepoAutomationService implements RepoAutomationService {
       readonly requireComplete?: boolean;
     },
   ): Promise<RepoAutomationExecutionResult> {
-    const args = buildValidateOrchestrationArtifactsArgs(input);
-
-    return this.executeScript({
-      tool: "validate_orchestration_artifacts",
-      runtimeKind: "python",
-      bundledRelativePath:
-        "resources/templates/validate_orchestration_artifacts.py",
-      workspaceRoot: input.workspaceRoot,
-      invocationId: input.invocationId ?? "validate_orchestration_artifacts",
-      args,
-      summary: `Validated ${input.artifactType} artifact at '${input.artifactPath}'.`,
-    });
+    return this.executeScript(
+      buildValidateOrchestrationArtifactsOptions(input),
+    );
   }
 
   private async executeScript(
